requirements.txt
flask==3.0.3
ultralytics==8.3.33
pillow==10.4.0
torch>=2.0.0
numpy>=1.23.0
